Remove `NoProviderStorage` as special case that is only necessary in bootstrap node
