+++ crates/subspace-networking/src/lib.rs
@@ -42,8 +42,8 @@ pub use crate::node::{
 pub use crate::node_runner::NodeRunner;
 pub use behavior::custom_record_store::{
     CustomRecordStore, LimitedSizeProviderStorageWrapper, LimitedSizeRecordStorageWrapper,
-    MemoryProviderStorage, NoProviderStorage, NoRecordStorage, ParityDbProviderStorage,
-    ParityDbRecordStorage, ProviderStorage, RecordStorage,
+    MemoryProviderStorage, NoRecordStorage, ParityDbProviderStorage, ParityDbRecordStorage,
+    ProviderStorage, RecordStorage,
 };
 pub use create::{create, peer_id, Config, CreationError, RelayMode};
 pub use libp2p;
